mudando para long int

+++ src/hash.h
@@ -6,7 +6,7 @@
 #include<stdbool.h>
 
 struct registro {
-  int chave;
+  long int chave;
   int dado;
 };
 
@@ -14,15 +14,15 @@ struct registro {
 typedef struct registro TipoElemento;
 typedef struct hash Hash;
 
-Hash* hash_criar(int tamanho);
-int   hash_tamanho(Hash *ha);
+Hash* hash_criar(long int tamanho);
+long int hash_tamanho(Hash *ha);
 void  hash_destruir(Hash** enderecoHash);
 bool  hash_inserir(Hash *ha, TipoElemento *elemento);
-bool  hash_remover(Hash *h, int chave, TipoElemento **elemento);
+bool  hash_remover(Hash *h, long int chave, TipoElemento **elemento);
 bool  hash_cheio(Hash *ha);
-bool  hash_getElemento(Hash *ha, int chave);
+bool  hash_getElemento(Hash *ha, long int chave);
 bool  hash_vazio(Hash *ha);
 void  hash_imprimir(Hash *ha);
-int   hash_buscar(Hash *h, int chave);
+int   hash_buscar(Hash *h, long int chave);
 
 #endif
